Move tests to separate Spec
