+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/TransformSpec.groovy
@@ -1,8 +1,6 @@
 package com.blackbuild.groovy.configdsl.transform
 
 import org.codehaus.groovy.control.MultipleCompilationErrorsException
-import org.junit.Rule
-import org.junit.rules.TemporaryFolder
 import spock.lang.Ignore
 import spock.lang.Issue
 
@@ -10,8 +8,6 @@ import java.lang.reflect.Method
 
 @SuppressWarnings("GroovyAssignabilityCheck")
 class TransformSpec extends AbstractDSLSpec {
-
-    @Rule TemporaryFolder temp = new TemporaryFolder()
 
     def "apply method is created"() {
         when:
@@ -325,190 +321,6 @@ class TransformSpec extends AbstractDSLSpec {
 
         then:
         noExceptionThrown()
-    }
-
-    def "convenience factory from script class"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                String value
-            }
-        ''')
-
-        def scriptClass = createSecondaryClass('''
-            import pk.Foo
-
-            Foo.create {
-                value "bla"
-            }
-        ''')
-
-        when:
-        instance = clazz.createFromScript(scriptClass)
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.value == "bla"
-    }
-
-    def "convenience factory from String"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                String value
-            }
-        ''')
-
-        def configText = '''
-            value "bla"
-        '''
-
-        when:
-        instance = clazz.createFromSnippet(configText)
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.value == "bla"
-    }
-
-    def "convenience factory with static methods"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                String value
-                Bar bar
-            }
-
-            @DSL
-            class Bar {
-                String bValue
-            }
-        ''')
-
-        def configText = '''
-            pk.Bar.createTemplate {
-                bValue "default"
-            }
-            value "bla"
-
-            bar {}
-
-        '''
-
-        when:
-        instance = clazz.createFromSnippet(configText)
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.value == "bla"
-        instance.bar.bValue == "default"
-    }
-
-    def "convenience factory from String with keyed object"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                @Key String name
-                String value
-            }
-        ''')
-
-        def configText = '''
-            value "bla"
-        '''
-
-        when:
-        instance = clazz.createFromSnippet("blub", configText)
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.name == "blub"
-        instance.value == "bla"
-    }
-
-    def "convenience factory from file"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                String value
-            }
-        ''')
-        File src = temp.newFile("blub.config")
-        src.text = '''
-            value "bla"
-        '''
-
-        when:
-        instance = clazz.createFromSnippet(src)
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.value == "bla"
-    }
-
-    def "keyed convenience factory from file"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                @Key String name
-                String value
-            }
-        ''')
-        File src = temp.newFile("blub.config")
-        src.text = '''
-            value "bla"
-        '''
-
-        when:
-        instance = clazz.createFromSnippet(src)
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.value == "bla"
-        instance.name == "blub"
-    }
-
-    def "keyed convenience factory from URL"() {
-        given:
-        createClass('''
-            package pk
-
-            @DSL
-            class Foo {
-                @Key String name
-                String value
-            }
-        ''')
-        File src = temp.newFile("blub.config")
-        src.text = '''
-            value "bla"
-        '''
-
-        when:
-        instance = clazz.createFromSnippet(src.toURI().toURL())
-
-        then:
-        instance.class.name == "pk.Foo"
-        instance.value == "bla"
-        instance.name == "blub"
     }
 
     def "constructor is created for keyed object"() {
